changed mc struct vars to float

+++ projects/motor_controller/inc/mc_data.h
@@ -11,9 +11,9 @@ typedef struct MotorData {
   uint8_t mc_limit_bitset;
   uint8_t mc_error_bitset;
   // updated every second
-  uint32_t motor_temp_c;
-  uint32_t heatsink_temp_c;
-  uint32_t dsp_temp_c;
+  float motor_temp_c;
+  float heatsink_temp_c;
+  float dsp_temp_c;
 } MotorData;
 
 MotorData left_motor;
